api/requirements.txt
# Framework API
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de données
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Cache & Messaging
redis==5.0.1
kafka-python==2.0.2

# Storage
minio==7.2.0

# Data processing
pandas==2.1.3
numpy==1.26.2

# Machine Learning (pour les prédictions)
scikit-learn==1.3.2
joblib==1.3.2

# Utilitaires
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.1

# Logging & Monitoring
prometheus-client==0.19.0
python-json-logger==2.0.7

# Authentification (optionnel)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4